Improve the performance of the `render-perf-tiles` test by using a singleton
color object rather than creating a new color each time. On my machine this
took the FPS of that test from 26FPS to 39FPS. A 50% improvement


The `Tileset#render` indentation was broken, so that's why the diff is a bit big.

The two major changes are:
  - Do not create a new color object for each vertices render (this occurs for each tile in a tileset)
  - Use an if/else for figuring out if vertices needs to rotate or not, calling a method and creating a new array object can have a large effect when it's called 10 000s / 100 000s of times per second