@mariovisic 2022-02-06
Here's a diff without whitespace (if that's easier to read): https://github.com/ruby2d/ruby2d/pull/231/files?w=1